Renamed to impl
