+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/gcc/impl.py
@@ -1,4 +1,4 @@
-import build_system.cmd.compiler.host.get_info.version.gnu.fetch_gnu_compiler_version
+import build_system.cmd.compiler.host.get_info.version.gnu.impl
 import build_system.compiler.family
 import build_system.compiler.version
