+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/msvc/impl.py
